refactor : 컴포넌트 스타일 파일명 변경 (styled -> style)
