+++ src/components/Common/BigTitle/style.tsx
