+++ src/pages/login/password_reset/style.tsx
